lvconvert: get new type from unmodified arg
The code that determines which operation is being done
needs to check the actual new type that was specified
by the command.  Checking lp->type_str won't work because
that value is changed throughout the code to trick other
parts of the code.  This means the correct operation is
not always identified.